--- components/form/demo/form-native.html
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<meta charset="UTF-8">
-	<link rel="stylesheet" href="../../demo/styles.css" type="text/css">
-	<script type="module">
-		import '../../demo/demo-page.js';
-		import './form-native-demo.js';
-	</script>
-</head>
-
-<body unresolved>
-
-	<d2l-demo-page page-title="d2l-form-native">
-
-		<h2>Basic</h2>
-		<d2l-demo-snippet>
-			<template>
-				<d2l-form-native-demo></d2l-form-native-demo>
-			</template>
-		</d2l-demo-snippet>
-
-	</d2l-demo-page>
-
-</body>
-
-</html>
