+++ components/form/docs/form.md
@@ -1,4 +1,4 @@
-# Form Components
+# Form [d2l-form]
 
 <!-- docs: demo -->
 ```html
@@ -34,12 +34,6 @@
 </d2l-form>
 ```
 
-There are two form components that can be used with our custom elements: `d2l-form` and `d2l-form-native`. These are useful in the following scenarios:
-- `d2l-form`: when submitting form data via your own API calls OR when nesting multiple forms within each other
-- `d2l-form-native`: when emulating native form element submission
-
-## Form [d2l-form]
-
 The `d2l-form` component can be used to build sections containing interactive controls that are validated and submitted as a group.
 
 It differs from the native HTML `form` element in 4 ways:
@@ -47,8 +41,6 @@ It differs from the native HTML `form` element in 4 ways:
 1. Upon validation, it will display an error summary that contains error messages for any elements that failed validation.
 1. `d2l-form` elements can be nested. If a parent form is validated or submitted it will also trigger the corresponding action for descendent `d2l-form`s unless they explicitly opt-out using `no-nesting`. This means that a `d2l-form` will only pass validation if it and all of its nested descendants pass validation.
 1. Submission is not handled directly by `d2l-form`. Instead, all form data will be aggregated and passed back to the caller via an event. The caller is then responsible for submitting the data.
-
-If you're looking to emulate native form element submission, `d2l-form-native` may be more appropriate.
 
 <!-- docs: demo code properties name:d2l-form sandboxTitle:'Form' autoSize:false display:block size:large -->
 ```html
@@ -174,68 +166,3 @@ In the above example, calling `submit` on form `#a` will cause forms `#a` and `#
 - `d2l-form#a` will be submitted because submit was called on it directly.
 - `d2l-form#b` will be submitted because it is nested directly within `#a`'s slot.
 - `d2l-form#root`, `d2l-form#c` and the `d2l-form` within the shadow root of `#d` will _**not**_ be submitted because they are ancestors of `#a` rather than descendants.
-
-## Native Form [d2l-form-native]
-
-The `d2l-form-native` component can be used to build sections containing interactive controls that are validated and submitted as a group.
-
-It differs from the native HTML `form` element in 2 ways:
-1. It supports custom form elements made using the [`FormElementMixin`](./form-element-mixin.md) in addition to native form elements like `input`, `select` and `textarea`.
-1. Upon validation, it will display an error summary that contains error messages for any elements that failed validation.
-
-If you're looking to submit form data via your own API calls or nest multiple forms within each other, `d2l-form` may be more appropriate.
-
-<!-- docs: demo code properties name:d2l-form-native sandboxTitle:'Native Form' autoSize:false display:block size:medium -->
-```html
-<script type="module">
-  import '@brightspace-ui/core/components/form/form-native.js';
-  import '@brightspace-ui/core/components/inputs/input-text.js';
-
-  const button = document.querySelector('button');
-  const form = document.querySelector('d2l-form-native');
-  button.addEventListener('click', () => {
-    form.submit();
-  });
-
-  function handleSubmission(e) {
-    const { formData } = e.detail;
-    const email = formData.get('email');
-    const pets = formData.get('pets');
-    console.log('Form submission data: email = ' + email + ', pets = ' + pets);
-  }
-  form.addEventListener('formdata', (e) => handleSubmission(e));
-</script>
-<d2l-form-native>
-  <d2l-input-text required label="Email" name="email" type="email"></d2l-input-text>
-  <select class="d2l-input-select" name="pets" required>
-    <option value="">--Please choose an option--</option>
-    <option value="porpoise">Porpoise</option>
-    <option value="house hippo">House Hippo</option>
-    <option value="spiker monkey">Spider Monkey</option>
-    <option value="capybara">Capybara</option>
-  </select>
-  <button name="action" value="save" type="submit">Save</button>
-</d2l-form-native>
-```
-
-<!-- docs: start hidden content -->
-### Properties
-
-| Property | Type | Description |
-|---|---|---|
-| `action` | String | The URL that processes the form submission. |
-| `enctype` | default: `"application/x-www-form-urlencoded"`<br>`"multipart/form-data"`<br>`"text/plain"` | If the value of the method attribute is post, enctype is the MIME type of the form submission. |
-| `method` | default: `"get"`<br>`"post"` | The URL that processes the form submission. |
-| `target` | default: `"_self"`<br>`"_blank"`<br>`"_parent"`<br>`"_top"` | Indicates where to display the response after submitting the form. |
-| `track-changes` | Boolean, default: `false` | Indicates that the form should interrupt and warn on navigation if the user has unsaved changes. |
-
-### Events
-- `submit`: Dispatched when the form is submitted. Cancelling this event will prevent form submission.
-- `formdata`: Dispatched after the entry list representing the form's data is constructed. This happens when the form is submitted just prior to submission. The form data can be obtained from the `detail`'s `formData` property.
-- `d2l-form-dirty`: Dispatched whenever any form element fires an `input` or `change` event. Can be used to track whether the form is dirty or not.
-<!-- docs: end hidden content -->
-
-### Methods
-- `submit()`: Submits the form to the server. This will first perform validation on all elements within the form. Submission will only happen if validation succeeds.
-- `requestSubmit(submitter)`: Requests that the form be submitted using the specified submit button and its corresponding configuration. A `button`'s value is only submitted if that button is both part of the form and the `submitter`.
-- `async validate()`: Validates the form without submitting even if validation succeeds. This returns a `Map` mapping from an element to the list of error messages associated with it.
