+++ index.html
@@ -90,13 +90,6 @@
 			</ul>
 		</li>
 		<li><a href="components/focus-trap/demo/focus-trap.html">d2l-focus-trap</a></li>
-		<li>
-			Forms
-			<ul>
-				<li><a href="components/form/demo/form.html">d2l-form</a></li>
-				<li><a href="components/form/demo/form-native.html">d2l-form-native</a></li>
-			</ul>
-		</li>
 		<li><a href="components/hierarchical-view/demo/hierarchical-view.html">d2l-hierarchical-view</a></li>
 		<li>
 			HtmlBlock
@@ -108,8 +101,9 @@
 		</li>
 		<li><a href="components/icons/demo/icon.html">d2l-icon</a></li>
 		<li>
-			Inputs
+			Inputs & Forms
 			<ul>
+				<li><a href="components/form/demo/form.html">d2l-form</a></li>
 				<li><a href="components/inputs/demo/input-checkbox.html">d2l-input-checkbox</a></li>
 				<li><a href="components/inputs/demo/input-color.html">d2l-input-color</a></li>
 				<li><a href="components/inputs/demo/input-date.html">d2l-input-date</a></li>
